projects/Agent-9_kl/requirements.txt
langchain-openai
langgraph
tavily-python
python-dotenv
requests
tweepy
pillow
openai
langsmith
google-genai
exa_py
crawl4ai
streamlit
fastapi
uvicorn[standard]
